commands: replace ESlint by Prettier
Command `npm run lint` was trying to find errors
in the code using ESlint. This command was also
executed as a pre-commit command.

As project is about to be rewritten in TypeScript,
linter is no longer needed but a formatter is.
Run command `npm run format` to format the code
using Prettier and `npm run format:check` to
verify code format. This last command is now the
one used as a pre-commit command.